+++ extensions/trello/src/searchTodos.tsx
@@ -1,6 +1,6 @@
 import { List } from "@raycast/api";
 import { useState } from "react";
-import { returnTodos } from "./utils/fetchTodos";
+import { trelloClient } from "./utils/trelloClient";
 import { TrelloFetchResponse } from "./trelloResponse.model";
 import { TodoListItem } from "./TrelloListItem";
 
@@ -10,17 +10,16 @@ export default function PackageList() {
 
   const onSearchTextChange = async (text: string) => {
     setLoading(true);
-    await returnTodos(text.replace(/\s/g, "+")).then((response) => {
-      setTodos(response);
-      setLoading(false);
-    });
+    const response = await trelloClient.searchCards(text);
+    setTodos(response);
+    setLoading(false);
   };
 
   return (
     <List isLoading={loading} searchBarPlaceholder={`Search todos`} onSearchTextChange={onSearchTextChange} throttle>
       {results?.length
         ? results.map((result) => {
-            return <TodoListItem key={result.id} result={result} />;
+            return <TodoListItem key={result.id} card={result} />;
           })
         : null}
     </List>